Add Ruff configuration
As repo-template, except for specifying Python version:

https://github.com/opensafely-core/repo-template/blob/e3d84fe3981d7f3df8f3cfa3c3a0a72e33cea5f2/pyproject.toml

Let's keep this in its own file for now, as we are not yet using
`pyproject.toml` in this project.

This also uses a few extra ignore options than we would normally specify
for new projects. We could remove these in future, but it seems less
disruptive to leave the affected code as is for now.